feat(staking): Tweak fonts
